Enable C# 8.0 nullability for the CLI, Android and iOS projects, improve help command behaviour (#145)
This enables the nullability and resolves all warnings connected to that in the projects.

There were a couple of changes I had to do (arguments re-work) in order to achieve this and some I did to improve working with the code:
- Moved option definition to Argument classes. This makes the code easier to work with because the option definitions are together with the properties that they fill and their validation.
- Improved the help command so that you can run `help [command] [subcommand]` which was not possible before. Now it's easier to explore the tool using calling `help xy`. It now also works properly even when you mess up the subcommand name.
- Moved the Darwin class to the iOS project and kept the CLI project strictly CLI.
- Dedicate the `ios test t=` argument to the `targets` option. It was originally assigned to `timeout` but `timeout` is optional and `targets` not. This change allows to call `ios test` with all required args using shortcuts (`-o -a -t`)

Fixes: #16

+++ src/Microsoft.DotNet.XHarness.CLI/Commands/iOS/iOSTestCommand.cs
@@ -4,7 +4,6 @@
 
 using System;
 using System.IO;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.DotNet.XHarness.CLI.CommandArguments;
@@ -17,7 +16,6 @@ using Microsoft.DotNet.XHarness.iOS.Shared.Listeners;
 using Microsoft.DotNet.XHarness.iOS.Shared.Logging;
 using Microsoft.DotNet.XHarness.iOS.Shared.Utilities;
 using Microsoft.Extensions.Logging;
-using Mono.Options;
 
 namespace Microsoft.DotNet.XHarness.CLI.Commands.iOS
 {
@@ -29,42 +27,11 @@ namespace Microsoft.DotNet.XHarness.CLI.Commands.iOS
         private readonly iOSTestCommandArguments _arguments = new iOSTestCommandArguments();
         private readonly ErrorKnowledgeBase _errorKnowledgeBase = new ErrorKnowledgeBase();
 
-        protected override ITestCommandArguments TestArguments => _arguments;
+        protected override string CommandUsage { get; } = "ios test [OPTIONS]";
+        protected override string CommandDescription { get; } = "Packaging command that will create a iOS/tvOS/watchOS or macOS application that can be used to run NUnit or XUnit-based test dlls";
+        protected override TestCommandArguments TestArguments => _arguments;
 
-        public iOSTestCommand() : base()
-        {
-            Options = new OptionSet()
-            {
-                "usage: ios test [OPTIONS]",
-                "",
-                "Packaging command that will create a iOS/tvOS/watchOS or macOS application that can be used to run NUnit or XUnit-based test dlls",
-                { "xcode=", "Path where Xcode is installed", v => _arguments.XcodeRoot = v},
-                { "mlaunch=", "Path to the mlaunch binary", v => _arguments.MlaunchPath = v},
-                { "device-name=", "Name of a specific device, if you wish to target one", v => _arguments.DeviceName = v},
-                { "communication-channel=", $"The communication channel to use to communicate with the default. Can be {CommunicationChannel.Network} and {CommunicationChannel.UsbTunnel}. Default is {CommunicationChannel.UsbTunnel}", v =>
-                    {
-                        if (Enum.TryParse(v, out CommunicationChannel channel))
-                        {
-                            _arguments.CommunicationChannel = channel;
-                        }
-                    }
-                },
-                { "launch-timeout=|lt=", "Time span, in seconds, to wait for the iOS app to start.", v => _arguments.LaunchTimeout = TimeSpan.FromSeconds(int.Parse(v))},
-                { "xml-jargon=|xj=", $"The xml format to be used in the unit test results. Can be {XmlResultJargon.TouchUnit} {XmlResultJargon.NUnitV2} {XmlResultJargon.NUnitV3} and {XmlResultJargon.xUnit}", v =>
-                    {
-                        // if we cannot parse it, set it as missing and the error will notify the issue
-                        _arguments.XmlResultJargon = Enum.TryParse(v, out XmlResultJargon jargon) ? jargon : XmlResultJargon.Missing;
-                    }
-                },
-            };
-
-            foreach (var option in CommonOptions)
-            {
-                Options.Add(option);
-            }
-        }
-
-        protected override async Task<ExitCode> InvokeInternal()
+        protected override async Task<ExitCode> InvokeInternal(ILogger logger)
         {
             var processManager = new ProcessManager(_arguments.XcodeRoot, _arguments.MlaunchPath);
             var deviceLoader = new HardwareDeviceLoader(processManager);
@@ -82,7 +49,7 @@ namespace Microsoft.DotNet.XHarness.CLI.Commands.iOS
                 var tunnelBore = (_arguments.CommunicationChannel == CommunicationChannel.UsbTunnel && !target.IsSimulator())
                     ? new TunnelBore(processManager)
                     : null;
-                var exitCodeForRun = await RunTest(target, logs, processManager, deviceLoader, simulatorLoader,
+                var exitCodeForRun = await RunTest(logger, target, logs, processManager, deviceLoader, simulatorLoader,
                     tunnelBore, cts.Token);
 
                 if (exitCodeForRun != ExitCode.SUCCESS)
@@ -94,21 +61,23 @@ namespace Microsoft.DotNet.XHarness.CLI.Commands.iOS
             return exitCode;
         }
 
-        private async Task<ExitCode> RunTest(TestTarget target,
+        private async Task<ExitCode> RunTest(
+            ILogger logger,
+            TestTarget target,
             Logs logs,
             ProcessManager processManager,
             IHardwareDeviceLoader deviceLoader,
             ISimulatorLoader simulatorLoader,
-            ITunnelBore tunnelBore,
+            ITunnelBore? tunnelBore,
             CancellationToken cancellationToken = default)
         {
-            _log.LogInformation($"Starting test for {target.AsString()}{ (_arguments.DeviceName != null ? " targeting " + _arguments.DeviceName : null) }..");
+            logger.LogInformation($"Starting test for {target.AsString()}{ (_arguments.DeviceName != null ? " targeting " + _arguments.DeviceName : null) }..");
 
             string mainLogFile = Path.Join(_arguments.OutputDirectory, $"run-{target}{(_arguments.DeviceName != null ? "-" + _arguments.DeviceName : null)}.log");
             ILog mainLog = logs.Create(mainLogFile, LogType.ExecutionLog.ToString(), true);
             int verbosity = GetMlaunchVerbosity(_arguments.Verbosity);
 
-            string deviceName = _arguments.DeviceName;
+            string? deviceName = _arguments.DeviceName;
 
             var appBundleInformationParser = new AppBundleInformationParser(processManager);
 
@@ -120,13 +89,13 @@ namespace Microsoft.DotNet.XHarness.CLI.Commands.iOS
             }
             catch (Exception e)
             {
-                _log.LogError($"Failed to get bundle information: {e.Message}");
+                logger.LogError($"Failed to get bundle information: {e.Message}");
                 return ExitCode.FAILED_TO_GET_BUNDLE_INFO;
             }
 
             if (!target.IsSimulator())
             {
-                _log.LogInformation($"Installing application '{appBundleInfo.AppName}' on " + (deviceName != null ? $" on device '{deviceName}'" : target.AsString()));
+                logger.LogInformation($"Installing application '{appBundleInfo.AppName}' on " + (deviceName != null ? $" on device '{deviceName}'" : target.AsString()));
 
                 var appInstaller = new AppInstaller(processManager, deviceLoader, mainLog, verbosity);
 
@@ -138,12 +107,12 @@ namespace Microsoft.DotNet.XHarness.CLI.Commands.iOS
                 }
                 catch (NoDeviceFoundException)
                 {
-                    _log.LogError($"Failed to find suitable device for target {target.AsString()}");
+                    logger.LogError($"Failed to find suitable device for target {target.AsString()}");
                     return ExitCode.DEVICE_NOT_FOUND;
                 }
                 catch (Exception e)
                 {
-                    _log.LogError($"Failed to install the app bundle:{Environment.NewLine}{e}");
+                    logger.LogError($"Failed to install the app bundle:{Environment.NewLine}{e}");
                     return ExitCode.PACKAGE_INSTALLATION_FAILURE;
                 }
 
@@ -153,38 +122,37 @@ namespace Microsoft.DotNet.XHarness.CLI.Commands.iOS
                     // the failure reason
                     if (_errorKnowledgeBase.IsKnownInstallIssue(mainLog, out var errorMessage))
                     {
-                        _log.LogError(
-                            $"Failed to install the app bundle (exit code={result.ExitCode}): {errorMessage}");
+                        logger.LogError($"Failed to install the app bundle (exit code={result.ExitCode}): {errorMessage}");
                     }
                     else
                     {
-                        _log.LogError($"Failed to install the app bundle (exit code={result.ExitCode})");
+                        logger.LogError($"Failed to install the app bundle (exit code={result.ExitCode})");
                     }
 
                     return ExitCode.PACKAGE_INSTALLATION_FAILURE;
                 }
 
-                _log.LogInformation($"Application '{appBundleInfo.AppName}' was installed successfully on device '{deviceName}'");
+                logger.LogInformation($"Application '{appBundleInfo.AppName}' was installed successfully on device '{deviceName}'");
             }
 
-            _log.LogInformation($"Starting application '{appBundleInfo.AppName}' on " + (deviceName != null ? $"device '{deviceName}'" : target.AsString()));
+            logger.LogInformation($"Starting application '{appBundleInfo.AppName}' on " + (deviceName != null ? $"device '{deviceName}'" : target.AsString()));
+
+            var appRunner = new AppRunner(
+                processManager,
+                deviceLoader,
+                simulatorLoader,
+                new SimpleListenerFactory(tunnelBore),
+                new CrashSnapshotReporterFactory(processManager),
+                new CaptureLogFactory(),
+                new DeviceLogCapturerFactory(processManager),
+                new TestReporterFactory(processManager),
+                mainLog,
+                logs,
+                new Helpers(),
+                useXmlOutput: true); // the cli ALWAYS will get the output as xml
 
             try
             {
-                var appRunner = new AppRunner(
-                    processManager,
-                    deviceLoader,
-                    simulatorLoader,
-                    new SimpleListenerFactory(tunnelBore),
-                    new CrashSnapshotReporterFactory(processManager),
-                    new CaptureLogFactory(),
-                    new DeviceLogCapturerFactory(processManager),
-                    new TestReporterFactory(processManager),
-                    mainLog,
-                    logs,
-                    new Helpers(),
-                    useXmlOutput: true); // the cli ALWAYS will get the output as xml
-
                 string resultMessage;
                 TestExecutingResult testResult;
                 (deviceName, testResult, resultMessage) = await appRunner.RunApp(appBundleInfo,
@@ -199,14 +167,14 @@ namespace Microsoft.DotNet.XHarness.CLI.Commands.iOS
                 switch (testResult)
                 {
                     case TestExecutingResult.Succeeded:
-                        _log.LogInformation($"Application finished the test run successfully");
-                        _log.LogInformation(resultMessage);
+                        logger.LogInformation($"Application finished the test run successfully");
+                        logger.LogInformation(resultMessage);
 
                         return ExitCode.SUCCESS;
 
                     case TestExecutingResult.Failed:
-                        _log.LogInformation($"Application finished the test run successfully with some failed tests");
-                        _log.LogInformation(resultMessage);
+                        logger.LogInformation($"Application finished the test run successfully with some failed tests");
+                        logger.LogInformation(resultMessage);
 
                         return ExitCode.TESTS_FAILED;
 
@@ -214,19 +182,19 @@ namespace Microsoft.DotNet.XHarness.CLI.Commands.iOS
 
                         if (resultMessage != null)
                         {
-                            _log.LogError($"Application run crashed:{Environment.NewLine}" +
+                            logger.LogError($"Application run crashed:{Environment.NewLine}" +
                                 $"{resultMessage}{Environment.NewLine}{Environment.NewLine}" +
                                 $"Check logs for more information.");
                         }
                         else
                         {
-                            _log.LogError($"Application run crashed. Check logs for more information");
+                            logger.LogError($"Application run crashed. Check logs for more information");
                         }
 
                         return ExitCode.APP_CRASH;
 
                     case TestExecutingResult.TimedOut:
-                        _log.LogWarning($"Application run timed out");
+                        logger.LogWarning($"Application run timed out");
 
                         return ExitCode.TIMED_OUT;
 
@@ -234,13 +202,13 @@ namespace Microsoft.DotNet.XHarness.CLI.Commands.iOS
 
                         if (resultMessage != null)
                         {
-                            _log.LogError($"Application run ended in an unexpected way: '{testResult}'{Environment.NewLine}" +
+                            logger.LogError($"Application run ended in an unexpected way: '{testResult}'{Environment.NewLine}" +
                                 $"{resultMessage}{Environment.NewLine}{Environment.NewLine}" +
                                 $"Check logs for more information.");
                         }
                         else
                         {
-                            _log.LogError($"Application run ended in an unexpected way: '{testResult}'. Check logs for more information");
+                            logger.LogError($"Application run ended in an unexpected way: '{testResult}'. Check logs for more information");
                         }
 
                         return ExitCode.GENERAL_FAILURE;
@@ -248,37 +216,37 @@ namespace Microsoft.DotNet.XHarness.CLI.Commands.iOS
             }
             catch (NoDeviceFoundException)
             {
-                _log.LogError($"Failed to find suitable device for target {target.AsString()}");
+                logger.LogError($"Failed to find suitable device for target {target.AsString()}");
                 return ExitCode.DEVICE_NOT_FOUND;
             }
             catch (Exception e)
             {
                 if (_errorKnowledgeBase.IsKnownTestIssue(mainLog, out var failureMessage))
                 {
-                    _log.LogError($"Application run failed:{Environment.NewLine}{failureMessage}");
+                    logger.LogError($"Application run failed:{Environment.NewLine}{failureMessage}");
                 }
                 else
                 {
-                    _log.LogError($"Application run failed:{Environment.NewLine}{e}");
+                    logger.LogError($"Application run failed:{Environment.NewLine}{e}");
                 }
 
                 return ExitCode.APP_CRASH;
             }
             finally
             {
-                if (!target.IsSimulator())
+                if (!target.IsSimulator() && deviceName != null)
                 {
-                    _log.LogInformation($"Uninstalling the application '{appBundleInfo.AppName}' from device '{deviceName}'");
+                    logger.LogInformation($"Uninstalling the application '{appBundleInfo.AppName}' from device '{deviceName}'");
 
                     var appUninstaller = new AppUninstaller(processManager, mainLog, verbosity);
                     var uninstallResult = await appUninstaller.UninstallApp(deviceName, appBundleInfo.BundleIdentifier, cancellationToken);
                     if (!uninstallResult.Succeeded)
                     {
-                        _log.LogError($"Failed to uninstall the app bundle with exit code: {uninstallResult.ExitCode}");
+                        logger.LogError($"Failed to uninstall the app bundle with exit code: {uninstallResult.ExitCode}");
                     }
                     else
                     {
-                        _log.LogInformation($"Application '{appBundleInfo.AppName}' was uninstalled successfully");
+                        logger.LogInformation($"Application '{appBundleInfo.AppName}' was uninstalled successfully");
                     }
                 }
             }
